feat(racingcar): accept and check user input
Function initGame: requests user input for names of cars and number of rounds. Evaluates user input: if names of cars is empty or length of car name is bigger than 5 throws IllegalArgumentException.
If number of rounds is less than one throws IllegalArgumentException. If couldn't parse number of rounds, throws NumberFormatException( : IllegalArgumentException)
Input is used to fill the properties of Game class object.